♻️ refactor: change deploy request page
